Centralize pipeline building logics in Buildkite::Builder::Pipeline (#37)
* Centralize pipeline building logics in Buildkite::Builder::Pipeline

* Fix specs

* Fix all

* Bump version with changelog

* Add missing file
