+++ lib/buildkite/pipelines.rb
@@ -6,7 +6,6 @@ module Buildkite
     autoload :Attributes, File.expand_path('pipelines/attributes', __dir__)
     autoload :Command, File.expand_path('pipelines/command', __dir__)
     autoload :Helpers, File.expand_path('pipelines/helpers', __dir__)
-    autoload :Pipeline, File.expand_path('pipelines/pipeline', __dir__)
     autoload :Plugin, File.expand_path('pipelines/plugin', __dir__)
     autoload :StepContext, File.expand_path('pipelines/step_context', __dir__)
     autoload :Steps, File.expand_path('pipelines/steps', __dir__)
