+++ spec/buildkite/builder/processors/abstract_spec.rb
@@ -3,8 +3,8 @@
 require 'logger'
 
 RSpec.describe Buildkite::Builder::Processors::Abstract do
-  let(:context) do
-    instance_double(Buildkite::Builder::Context, logger: Logger.new)
+  let(:pipeline) do
+    instance_double(Buildkite::Builder::Pipeline, logger: Logger.new)
   end
 
   describe '.process' do
@@ -22,12 +22,12 @@ RSpec.describe Buildkite::Builder::Processors::Abstract do
     let(:abstract_processor) { described_class }
 
     it 'raises an error' do
-      expect { abstract_processor.process(context) }.to raise_error(NotImplementedError)
+      expect { abstract_processor.process(pipeline) }.to raise_error(NotImplementedError)
     end
 
     context 'with an implemented processor' do
       it 'processes' do
-        expect(foo_processor.process(context)).to be_truthy
+        expect(foo_processor.process(pipeline)).to be_truthy
       end
     end
   end
