+++ spec/fixtures/simple/.buildkite/pipeline.rb
@@ -1,0 +1,6 @@
+Buildkite::Builder.pipeline do
+  command do
+    label 'Simple step'
+    command :noop
+  end
+end
